split task in separate dir
